changes for new api
