+++ .gitignore
@@ -65,7 +65,7 @@ db.sqlite3-journal
 instance/
 .webassets-cache
 /migrations/
-
+migrations/
 # Scrapy stuff:
 .scrapy
